# docs/swagger.yaml
basePath: /
definitions:
  github_com_NStegura_gophermart_internal_app_gophermartapi_models.Balance:
    properties:
      current:
        type: number
      withdrawn:
        type: number
    type: object
  github_com_NStegura_gophermart_internal_app_gophermartapi_models.Order:
    properties:
      accrual:
        type: number
      number:
        example: "0"
        type: string
      status:
        type: string
      uploaded_at:
        type: string
    type: object
  github_com_NStegura_gophermart_internal_app_gophermartapi_models.User:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  github_com_NStegura_gophermart_internal_app_gophermartapi_models.WithdrawIn:
    properties:
      order:
        type: string
      sum:
        type: number
    type: object
  github_com_NStegura_gophermart_internal_app_gophermartapi_models.WithdrawOut:
    properties:
      order:
        type: string
      processed_at:
        type: string
      sum:
        type: number
    type: object
info:
  contact: {}
  description: This is a Gophermart server.
  title: Gophermart API
  version: "1.0"
paths:
  /api/user/balance:
    get:
      description: get user balance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.Balance'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get balance
      tags:
      - user
  /api/user/balance/withdraw:
    post:
      consumes:
      - application/json
      description: create user withdraw
      parameters:
      - description: User withdraw data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.WithdrawIn'
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "402":
          description: Payment Required
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create withdraw
      tags:
      - user
  /api/user/login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.User'
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Use this header in other endpoints
              type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /api/user/orders:
    get:
      description: get order list by user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.Order'
            type: array
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get order list
      tags:
      - user
    post:
      consumes:
      - text/plain
      description: '"register" user order'
      parameters:
      - description: Order id
        in: body
        name: string
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "202":
          description: Accepted
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create order
      tags:
      - user
  /api/user/register:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.User'
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Use this header in other endpoints
              type: string
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Register
      tags:
      - auth
  /api/user/withdrawals:
    get:
      description: get user withdraw list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_NStegura_gophermart_internal_app_gophermartapi_models.WithdrawOut'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get withdraw list
      tags:
      - user
  /ping:
    get:
      description: check service
      responses:
        "200":
          description: OK
      summary: Get ping
      tags:
      - tech
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"